+++ test/instance/values.test.js
@@ -48,7 +48,9 @@ describe(Support.getTestDialectTeaser('DAO'), function() {
           identifier: {type: DataTypes.STRING, primaryKey: true}
         });
 
-        var user = User.build();
+        var user = User.build({}, {
+          isNewRecord: false
+        });
 
         user.set({
           createdAt: new Date(2000, 1, 1),
@@ -66,7 +68,9 @@ describe(Support.getTestDialectTeaser('DAO'), function() {
           underscored: true
         });
 
-        var user = User.build();
+        var user = User.build({}, {
+          isNewRecord: false
+        });
 
         user.set({
           created_at: new Date(2000, 1, 1),
@@ -353,6 +357,24 @@ describe(Support.getTestDialectTeaser('DAO'), function() {
 
           expect(product.get('user', {plain: true}).$Model).not.to.be.ok;
           expect(product.get({plain: true}).user.$Model).not.to.be.ok;
+        });
+      });
+
+      describe('clone', function() {
+        it('should copy the values', function () {
+          var Product = this.sequelize.define('product', {
+            title: Sequelize.STRING
+          });
+
+          var product = Product.build({
+            id: 1,
+            title: 'Chair',
+          }, {raw: true});
+
+          var values = product.get({clone: true});
+          delete values.title;
+
+          expect(product.get({clone: true}).title).to.be.ok;
         });
       });
     });